+++ kubejs/client_scripts/mms_jei.js
@@ -1,0 +1,7 @@
+// priority: 1
+
+JEIEvents.addItems(event => {
+  event.add('minecraft:barrier')
+})
+
+console.info('mms_jei.js load completed')